cameracontrols.js: Add coordinateTransformation() support also for rotateX/Y/Z functions.
Add correction for rotating the camera. This approach is completely
different than the simple sin()*cos() correction that is possible
for translations (linear movement). It turns out, it is not possible
to figure out a similar correction for rotations. On the other hand
order of applying the rotations matter, and it turns out whichever
axis is rotated first, will look natural in the camera (which is
the same as the impress.js canvas div). Note that the canvas
applies the rotations in the opposite order than what is
specified for each element, so the axis to be rotated *first*
on the canvas, is actually *last* in the order field.

I found that changing the order of xyz in the "order" field
works smoothly when the rotation for an axis is 0. But when
it is non-zero, flipping the order will cause an ongoing
rotation to look like it completely changed direction all
of a sudden (which is precisely what it would do). Hence
fiddling with the order of rotation axes is only done if
the rotation for the axis is currently 0, and the axis isn't
already rotating. Once the order cannot be changed anymore,
the universe will simply rotate in some other direction than
facing the camera.

Introduced data-rotate-order and the related UI field in the
toolbar to allow for reordering of rotations. The corresponding
addition in impress.js itself is 9ab6b2b69a325fa088e89550d2b193286658f852

Ref: A blog post on how I learned to understand how CSS in general
and impress.js in particular apply rotation orders:
http://openlife.cc/blogs/2016/october/3d-rotations-css-and-impressjs

Ref: Completes 8178380b6c1489f4cac0a6265a48127864772669 which added
coordinateTransformation() corrections for cameracontrols navigation
in x, y, and z directions (i.e. translate).