Move tinymce integration out from the presentation itself, into src/plugins.
Note that the <script> element that includes tinymce itself is still in
each of the index.html files of the impress.js presentations. Hence, this
still only works only with a file in templates/*/index.html. Future work
will completely extract tinymce.js out from the html, so that even the
<script> tag itself is added by this tinymce integration plugin.