Add Electron framework.
As of this commit you can now do

    npm start

to launch the electron based app, choose File -> Open to
open templates/cube/index.html, edit the slides (but not
create or add slides) and the save it with File -> Save as.

Note: It won't work with any other impress presentation, as
tinymce.js and impressionist.[js|css] are still in the index.html
file hard coded. Also the saved file isn't really clean, since it
will save the tinymce toolbar, dialogs and the impressionist camera
controls. All of this will change in future commits of course.

So the cool part of this commit is simply that you can launch an
Electron app now.