Add Step tab to the toolbar, with controls to position the step.
The controls are exactly the same as the existing tab to move the
Camera (which is the root and canvas in impress.js terminology).
For this reason the code is not a new plugin, rather was added
into the camera.js plugin. I'm not too enthusiastic about this
entanglement, and look forward to refactoring this into several
smaller plugins in the future.

Note that the cameracontrols.js (the + O Z buttons) will still only
control the camera, they don't know about the steps yet. This is
the focus of following commits.