Enable the cameracontrols to also move the step / coordinates of Step toolbar tab.
If the "lock" checkbox is checked, using the + O Z cameracontrols will
move both the camera and the current step. In other words, the current
step doesn't move respective to current window, but everythin else does
move behind and around it.

If the "lock" checkbox is unchecked, then the cameracontrols will move
either the Camera or the Step, depending of which toolbar tab is active.
Is neither toolbar tab is visible, it will move the Camera.