Update to newest impress.js.
This update pulls in

d8f21f303e604374177d9c53edb1c381a66a5d2a
Add support for dynamic changes to steps

This is needed due to the recent new feature of moving a step.
impress.js core needs to become aware of the changed coordinates,
otherwise it would just navigate to the original position of the step.

This impress.js enhancement will also make it easy to implement addition
or removal of new slides. This is next on the todo list!